requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pinecone-client>=3.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.1
transformers>=4.30.0
torch>=2.0.0
python-serpapi>=0.1.5
pypdf>=3.17.0
unstructured>=0.12.0
tf-keras>=1.0.0